requirements.txt
pandas~=1.2.3
torch~=1.9.0
Pillow~=8.1.0
torchvision~=0.10.0
scikit-learn~=0.24.2
matplotlib~=3.3.4
tqdm~=4.57.0